reindex fixed

+++ src/Enhavo/Bundle/SearchBundle/Index/IndexEngine.php
@@ -505,40 +505,42 @@ class IndexEngine implements IndexEngineInterface {
     function searchUpdateTotals() {
 
         //get all of the saved words from seach_dirty
-        foreach ($this->searchDirty() as $word => $dummy) {
+        if($this->searchDirty() != null) {
+            foreach ($this->searchDirty() as $word => $dummy) {
 
-            // Get total count for the word
-            $searchIndexRepository = $this->em->getRepository('EnhavoSearchBundle:Index');
-            $total = $searchIndexRepository->sumScoresOfWord($word);
+                // Get total count for the word
+                $searchIndexRepository = $this->em->getRepository('EnhavoSearchBundle:Index');
+                $total = $searchIndexRepository->sumScoresOfWord($word);
 
-            //get the total count: Normalization according Zipf's law --> the word's value to the search index is inversely proportionate to its overall frequency therein
-            $total = log10(1 + 1/(max(1, current($total))));
+                //get the total count: Normalization according Zipf's law --> the word's value to the search index is inversely proportionate to its overall frequency therein
+                $total = log10(1 + 1 / (max(1, current($total))));
 
-            //save new score
-            $searchTotalRepository =  $this->em->getRepository('EnhavoSearchBundle:Total');
-            $currentWord = $searchTotalRepository->findBy(array('word' => $word));
+                //save new score
+                $searchTotalRepository = $this->em->getRepository('EnhavoSearchBundle:Total');
+                $currentWord = $searchTotalRepository->findBy(array('word' => $word));
 
-            foreach ($currentWord as $cWord) {
-                //if the word is already stored in search_total -> remove it and store it with the new score
-                $this->em->remove($cWord);
+                foreach ($currentWord as $cWord) {
+                    //if the word is already stored in search_total -> remove it and store it with the new score
+                    $this->em->remove($cWord);
+                }
+
+                $newTotal = new Total();
+                $newTotal->setWord($word);
+                $newTotal->setCount($total);
+                $this->em->persist($newTotal);
+                $this->em->flush();
             }
 
-            $newTotal = new Total();
-            $newTotal->setWord($word);
-            $newTotal->setCount($total);
-            $this->em->persist($newTotal);
-            $this->em->flush();
-        }
-
-        //remove words that are removed vom search_index but are still in search_total
-        $searchTotalRepository = $this->em->getRepository('EnhavoSearchBundle:Total');
-        $wordsToRemove = $searchTotalRepository->getWordsToRemove();
-        foreach ($wordsToRemove as $word) {
-            $currentWordsToRemove = $searchTotalRepository->findBy(array('word' => $word['realword']));
-            if($currentWordsToRemove != null){
-                foreach($currentWordsToRemove as $currentWordToRemove) {
-                    $this->em->remove($currentWordToRemove);
-                    $this->em->flush();
+            //remove words that are removed vom search_index but are still in search_total
+            $searchTotalRepository = $this->em->getRepository('EnhavoSearchBundle:Total');
+            $wordsToRemove = $searchTotalRepository->getWordsToRemove();
+            foreach ($wordsToRemove as $word) {
+                $currentWordsToRemove = $searchTotalRepository->findBy(array('word' => $word['realword']));
+                if ($currentWordsToRemove != null) {
+                    foreach ($currentWordsToRemove as $currentWordToRemove) {
+                        $this->em->remove($currentWordToRemove);
+                        $this->em->flush();
+                    }
                 }
             }
         }
